+++ perf/_runner.py
@@ -492,8 +492,8 @@ class Runner:
         task.inner_loops = inner_loops
         return self._main(task)
 
-    def timeit(self, name, stmt, setup="pass", inner_loops=None,
-               duplicate=None, metadata=None, globals=None):
+    def timeit(self, name, stmt, setup="pass", teardown="pass",
+               inner_loops=None, duplicate=None, metadata=None, globals=None):
 
         if not self._check_worker_task():
             return None
@@ -502,6 +502,7 @@ class Runner:
         from perf._timeit import bench_timeit
         return bench_timeit(self, name, stmt,
                             setup=setup,
+                            teardown=teardown,
                             inner_loops=inner_loops,
                             duplicate=duplicate,
                             func_metadata=metadata,